% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{plot19ts}
\alias{plot19ts}
\title{This is a function for producing Covid-19 time series plot}
\usage{
plot19ts(sel_cou = NULL, covid19df = NULL, yvar = "hosp")
}
\description{
This is a function for producing Covid-19 time series plot
}
